+++ backend/substrapp/signals/model/post_delete.py
@@ -1,0 +1,7 @@
+from substrapp.storages.minio import MinioStorage
+
+
+def model_post_delete(sender, instance, **kwargs):
+    # delete file from MinioStorage
+    if isinstance(instance.file.storage, MinioStorage):
+        instance.file.storage.delete(str(instance.file.name))